requirements.txt
Django==1.9.1
django-filters==0.2.1
djangorestframework==3.3.2
mysqlclient==1.3.7
wheel==0.24.0
